+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Added
+
+- Export `TokenRequestIncomingMessage` type to make `/token` related hooks easier to consume (reported in [#355](https://github.com/axa-group/oauth2-mock-server/issues/355) by [kikisaeba](https://github.com/kikisaeba))
+
 ## [8.1.0](https://github.com/axa-group/oauth2-mock-server/compare/v8.0.1...v8.1.0) — 2025-06-06
 
 ### Added